+++ CHANGELOG.md
@@ -16,7 +16,7 @@ FIXES
 
 * `list_runs` was returning empty response
 * List runs was returning empty response [275](https://github.com/hashicorp/terraform-mcp-server/pull/275)
-* Generate session id when `MCP_SESSION_MODE` is `stateless`
+* Fix sessions handling in stateless and load balanced environments
 
 ## 0.4.0